Tag: V00-04-05 2013-05-22 Andy Salnikov - update ImpSamples element type
git-svn-id: https://pswww.slac.stanford.edu/svn-readonly/psdmrepo@6353 b967ad99-d558-0410-b138-e0f6c56caec7

+++ H5DataTypes/trunk/include/ImpElementV1.h
@@ -52,7 +52,7 @@ public:
   static hdf5pp::Type native_type() ;
 
 private:
-  uint32_t channels[4]; 
+  uint16_t channels[4];
 };
 
 //
